+++ bin/src/run/loadConfigFile.ts
@@ -1,4 +1,3 @@
-import buble from 'rollup-plugin-buble';
 import path from 'path';
 import chalk from 'chalk';
 import * as rollup from 'rollup';
@@ -20,7 +19,6 @@ export default function loadConfigFile (configFile: string, silent = false): Pro
 				);
 			},
 			onwarn: warnings.add,
-			plugins: [buble({ objectAssign: 'Object.assign' })]
 		})
 		.then((bundle: OutputChunk) => {
 			if (!silent && warnings.count > 0) {